requirements.txt
pillow>=10.0.0
typer>=0.12.3
